+++ main.ts
@@ -9,13 +9,10 @@ function createWindow () {
     height: 600,
     backgroundColor: '#ffffff',
     icon: `file://${__dirname}/dist/assets/logo.png`,
-    frame: true
+    frame: false
   });
 
   win.loadURL(`file://${__dirname}/dist/index.html`)
-
-  //// uncomment below to open the DevTools.
-  // win.webContents.openDevTools()
 
   // Event when the window is closed.
   win.on('closed', function () {