+++ src/app/core/top-nav/top-nav.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { RemoteService } from 'src/app/electron-utils/remote.service';
+import { remote } from 'electron';
 
 @Component({
   selector: 'top-nav',
@@ -8,11 +9,26 @@ import { RemoteService } from 'src/app/electron-utils/remote.service';
 })
 export class TopNavComponent implements OnInit {
   remote = new RemoteService().remote;
+  win = this.remote.getCurrentWindow();
 
   constructor() {
   }
 
   ngOnInit() {
+  }
+
+  openDevTools() {
+    this.win.webContents.openDevTools();
+  }
+
+  minimize() {
+    var window = this.remote.getCurrentWindow();
+    window.minimize();
+  }
+
+  maximize() {
+    var window = this.remote.getCurrentWindow();
+    window.isMaximized() ? window.unmaximize() : window.maximize();
   }
 
   close() {